fix migration
