+++ deployment/.env.template
@@ -2,7 +2,6 @@
 DOMAIN=
 
 # app
-DB_HOST=
 DB_USER=
 DB_PASSWORD=
 DB_NAME=
