--- deployment/docker-compose.dev.yml
@@ -1,11 +1,0 @@
-services:
-  woben_app:
-    build:
-      context: ../
-      dockerfile: ./deployment/Dockerfile
-      args:
-        - DB_URL=postgresql://${DB_USER}:${DB_PASSWORD}@${DB_HOST}/${DB_NAME}?sslmode=require
-    container_name: woben_app
-    restart: always
-    ports:
-      - 3000:3000
