+++ platforms/app/src/lib/db/node-postgres-client.ts
@@ -1,11 +1,9 @@
 import { serverEnv } from "@/env";
-import pg from "pg";
 import { drizzle } from "drizzle-orm/node-postgres";
+import pg from "pg";
 
-export async function createNodePostgresClient() {
-	const client = new pg.Pool({
-		connectionString: serverEnv.DATABASE_URL,
-	});
+export function createNodePostgresClient() {
+	const client = new pg.Pool({ connectionString: serverEnv.DATABASE_URL });
 	const db = drizzle({ client });
 	return db;
 }
